Fully finished the project
Added a local storage for the cart and favourites items. Also build a function to toggle between the different pages (main menu, cart, favourites). I would say that, the project is almost done, there will always a few little details in my opinion at least, that can be changed and updated, but these are so small that it makes no difference on the functionality of the app. There may be more updates to it soon, though!